+++ cloudmanager/resource_netapp_cloudmanager_connector_azure.go
@@ -142,6 +142,11 @@ func resourceOCCMAzure() *schema.Resource {
 				Optional: true,
 				ForceNew: true,
 			},
+			"retries": {
+				Type:     schema.TypeInt,
+				Optional: true,
+				Default:  150,
+			},
 		},
 	}
 }
@@ -223,7 +228,7 @@ func resourceOCCMAzureCreate(d *schema.ResourceData, meta interface{}) error {
 		occmDetails.StorageAccount = o.(string)
 	}
 
-	res, err := client.createOCCMAzure(occmDetails, proxyCertificates, "")
+	res, err := client.createOCCMAzure(occmDetails, proxyCertificates, "", d.Get("retries").(int))
 	if err != nil {
 		log.Print("Error creating instance")
 		return err