+++ cloudmanager/resource_netapp_cloudmanager_connector_gcp.go
@@ -155,6 +155,11 @@ func resourceOCCMGCP() *schema.Resource {
 					Type: schema.TypeString,
 				},
 			},
+			"retries": {
+				Type:     schema.TypeInt,
+				Optional: true,
+				Default:  150,
+			},
 		},
 	}
 }
@@ -244,7 +249,7 @@ func resourceOCCMGCPCreate(d *schema.ResourceData, meta interface{}) error {
 		occmDetails.Tags = tags
 	}
 
-	res, err := client.deployGCPVM(occmDetails, proxyCertificates, "")
+	res, err := client.deployGCPVM(occmDetails, proxyCertificates, "", d.Get("retries").(int))
 	if err != nil {
 		log.Print("Error creating instance")
 		return err